Create Class Problem
Create a class, named Student, representing the student's details, as mentioned above, and store the data of a student. Create setter and getter functions for each element; that is, the class should at least have following functions:

1. get_age, set_age
2. get_first_name, set_first_name
3. get_last_name, set_last_name
4. get_standard, set_standard
Also, you have to create another method to_string() which returns the string consisting of the above elements, separated by a comma(,).
You can refer to stringstream for this.